requirements.txt
algoliasearch==1.20.0
algoliasearch-django==1.7.3
asgiref==3.3.1
boto3==1.16.56
botocore==1.19.63
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
google-api-core==1.26.0
google-auth==1.26.1
google-cloud-core==1.6.0
google-cloud-datastore==1.15.3
google-cloud-ndb==1.7.3
googleapis-common-protos==1.52.0
grpcio==1.35.0
idna==2.10
jmespath==0.10.0
packaging==20.9
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pymemcache==3.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-http-client==3.3.2
pytz==2021.1
redis==3.5.3
requests==2.25.1
rsa==4.7.1
s3transfer==0.3.6
sendgrid==6.5.0
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
stripe==2.56.0
urllib3==1.26.3
